Make okhttp tests ambivalent about TLS 1.x.
This is intended as a test-only change suitable for backporting
to historic CTS branches where vendors may have disabled TLS 1.x.

It does add the TLS 1.3 protocol and its ciphers to the relevant
enums, which ship in the platform but are not actually used
on Android.

All Android HttpsURLConnections use a predefined ConnectionSpec
defined in HttpsHandler which enables all cipher suites and
protocols, i.e. on Android, URL connections are intended to
pick up the correct parameters from their socket factory,
rather than using the ConnectionSpec API to set them.

tl;dr much of the changes are to tests for APIs which are not
reachable on Android, but which we should continue to test.

Bug: 288058606
Test: atest CtsLibcoreOkHttpTestcases (with and without TLS 1.0/1.1 disabled)
Change-Id: I57bd878f25a83763933a453dc6e18c349ed28028