Grant okhttp-norepackage for packages under /vendor.
We will introduce Firebase Cloud Firestore for RamdumpService by ag/30111643.
When we imported the offcial com.squareup.okio required by Cloud Firestore, we ran into a runtime issue:
Stream closed with status: Status{code=INTERNAL, description=Panic! This is a bug!, cause=java.lang.NoClassDefFoundError: Failed resolution of: Lokio/ByteString

We use this modified version intead to fix it.
RammdumpService is under /vendor, so we grant the permission for /vendor.

Bug: 369260803

Change-Id: I7ce0c447f7f4392485b8e7e27fc1933b7fb4abcc